refactor network internal auth helpers
